BLAS/legacylens/requirements.txt
chromadb>=0.4.0
voyageai>=0.2.0
python-dotenv>=1.0.0
anthropic>=0.18.0
rich>=13.0.0
